+++ src/basic/App.tsx
@@ -11,48 +11,9 @@ import {
   NotificationVariant,
   type Notification as NotificationType,
 } from "./entities/notification/types";
-import { getRemainingStock } from "./features/check-stock/libs";
 import { useCartStorage } from "./entities/cart/hooks/useCartStorage";
-
-interface ProductWithUI extends Product {
-  description?: string;
-  isRecommended?: boolean;
-}
-
-// 초기 데이터
-const initialProducts: ProductWithUI[] = [
-  {
-    id: "p1",
-    name: "상품1",
-    price: 10000,
-    stock: 20,
-    discounts: [
-      { quantity: 10, rate: 0.1 },
-      { quantity: 20, rate: 0.2 },
-    ],
-    description: "최고급 품질의 프리미엄 상품입니다.",
-  },
-  {
-    id: "p2",
-    name: "상품2",
-    price: 20000,
-    stock: 20,
-    discounts: [{ quantity: 10, rate: 0.15 }],
-    description: "다양한 기능을 갖춘 실용적인 상품입니다.",
-    isRecommended: true,
-  },
-  {
-    id: "p3",
-    name: "상품3",
-    price: 30000,
-    stock: 20,
-    discounts: [
-      { quantity: 10, rate: 0.2 },
-      { quantity: 30, rate: 0.25 },
-    ],
-    description: "대용량과 고성능을 자랑하는 상품입니다.",
-  },
-];
+import { useProductStorage } from "./entities/product/hooks/useProductStorage";
+import { getRemainingStock } from "./features/check-stock/libs";
 
 const initialCoupons: Coupon[] = [
   {
@@ -71,10 +32,7 @@ const initialCoupons: Coupon[] = [
 
 const App = () => {
   const search = useSearch();
-  const [products, setProducts] = useLocalStorageObject<ProductWithUI[]>(
-    "products",
-    initialProducts
-  );
+  const { products, setProducts } = useProductStorage();
   const { cart, setCart, totalItemCount } = useCartStorage();
   const [coupons, setCoupons] = useLocalStorageObject<Coupon[]>(
     "coupons",
